MHC-715 Use Bridge's "V1 Legacy Bundles" format for AppCore compatibility when uploading surveys
In Bridge's "V2 Generic format", all the survey answers are stored in one file, by default in answers.json file. In AppCore, the answer to each survey question is stored in its own file. To be compatible with this approach, Bridge SDK needs to use "V1 Legacy Bundles" format when uploading surveys.